delete component.

+++ Morty/Core/Utility/MFunction.h
@@ -81,7 +81,7 @@ bool ERASE_FIRST_VECTOR(std::vector<T>& vector, const T& value, const std::funct
 }
 
 template<typename T>
-uint32_t UNION_ORDER_PUSH_BACK_VECTOR(std::vector<T>& vector, const T& value, const std::function<bool(const T& a, const T& b)>& lessComp = std::less<T>(), const std::function<bool(const T& a, const T& b)>& equalComp = std::equal_to<T>())
+size_t UNION_ORDER_PUSH_BACK_VECTOR(std::vector<T>& vector, const T& value, const std::function<bool(const T& a, const T& b)>& lessComp = std::less<T>(), const std::function<bool(const T& a, const T& b)>& equalComp = std::equal_to<T>())
 {
 	if (vector.empty())
 	{
